Consider the config files as specified by the specs
The specs states that the config file can be in one of the following paths, in
the order of precedence:

- .devcontainer/devcontainer.json
- .devcontainer.json
- .devcontainer/<folder>/devcontainer.json (where <folder> is a sub-folder, one level deep)

This commit adds a utility function that can be used later to select the config
file. For now, it is used to check whether a container is needed or not.