Introduce --docker-path, call-engine-string-sync and make-cli-args
`devcontainer--docker-path' returns the correct path for the selected
method. It's using the tramp variables because that way we ensure we
are using the same engine command. This may mean we have to require
Tramp to properly read these variables. Docker by default for
backwards compat.

`devcontainer--call-engine-string-sync' is similar to
`shell-command-to-string' but we can make it specific to container
engines, it uses argument vectors rather than stringly shell commands,
and also our error handling is explicit.

`devcontainer--make-cli-args' is an argument builder for the
devcontainer CLI. This is to centralise the arguments passed to the CLI.