Change functions to use -engine instead of hard-coding docker
`devcontainer--determine-container-id-cmd' was called only here, so if
we put it in a lambda we can colocate it where it is used. Let-binding
a lambda is ugly though, could be changed if we wished to require cl-lib.

Make `devcontainer-image-id' engine-independent

Make `devcontainer-is-up' engine-independent

Make `devcontainer-up' engine-independent

Make `devcontainer-kill-container' engine-independent

Change functions to use -engine instead of hard-coding docker